+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.1.0 - 1.21.1
+- Spell Engine 1.2 Update
+- Smoldering Arrow, Frozen Shot, Fast Shot got reworked, they use the new SpellStash Feature
+- all arrow spells now have special models & travel particle visuals
+- Poison Immune Entites are now also immune to the Choking Gas Effect
+- changed the Choking Gas Spell a bit
+
 1.0.2 - 1.21.1
 - #3 Not compatible with 1.21 while it says it is on Modrinth #3 fix
 - forgot enchantment item tags for the armor sets